bigger matrix, data sending, refactoring

+++ tasks/gutyansky_a_matrix_column_sum/common/include/common.hpp
@@ -10,7 +10,7 @@
 namespace gutyansky_a_matrix_column_sum {
 
 using InType = Matrix;
-using OutType = std::vector<int64_t>;
+using OutType = std::vector<int32_t>;
 using TestType = std::string;
 using BaseTask = ppc::task::Task<InType, OutType>;
