+++ tasks/gutyansky_a_matrix_column_sum/common/include/matrix.hpp
@@ -10,7 +10,7 @@ namespace gutyansky_a_matrix_column_sum {
 struct Matrix {
   size_t rows;
   size_t cols;
-  std::vector<int64_t> data;
+  std::vector<int32_t> data;
 
   friend bool operator==(const Matrix &v_left, const Matrix &v_right) {
     return v_left.rows == v_right.rows && v_left.cols == v_right.cols && v_left.data == v_right.data;
